+++ Server/Controllers/FootballHistoryController.cs
@@ -49,26 +49,5 @@ namespace football_history.Controllers
 
             return league;
         }
-
-        public class LeagueTable
-        {
-            public string Competition { get; set; }
-            public string Season { get; set; }
-            public List<LeagueTableRow> LeagueTableRow { get; set; }
-        }
-
-        public class LeagueTableRow
-        {
-            public int Position { get; set; }
-            public string Team { get; set; }
-            public int Played { get; set; }
-            public int Won { get; set; }
-            public int Drawn { get; set; }
-            public int Lost { get; set; }
-            public int GoalsFor { get; set; }
-            public int GoalsAgainst { get; set; }
-            public int GoalDifference { get; set; }
-            public int Points { get; set; }
-        }
     }
 }
